+++ src/revlanguage/distributions/phylogenetics/tree/Dist_divDepYuleProcess.cpp
@@ -60,7 +60,7 @@ RevBayesCore::DiversityDependentPureBirthProcess* Dist_divDepYuleProcess::create
     // speciation rate
     RevBayesCore::TypedDagNode<double>* s       = static_cast<const RealPos &>( initialLambda->getRevObject() ).getDagNode();
     // extinction rate
-    RevBayesCore::TypedDagNode<int>* k          = static_cast<const Natural &>( capacity->getRevObject() ).getDagNode();
+    RevBayesCore::TypedDagNode<long>* k          = static_cast<const Natural &>( capacity->getRevObject() ).getDagNode();
     // condition
     const std::string& cond                     = static_cast<const RlString &>( condition->getRevObject() ).getValue();
     // taxon names
