+++ src/revlanguage/functions/internal/Func__mod.h
@@ -34,7 +34,7 @@ namespace RevLanguage {
         bool                                        isInternal(void) const { return true; }                     //!< Is this an internal function?
 
         // Regular functions
-        RevBayesCore::TypedFunction<int>*           createFunction(void) const;                                                     //!< Create internal function object
+        RevBayesCore::TypedFunction<long>*           createFunction(void) const;                                                     //!< Create internal function object
         const ArgumentRules&                        getArgumentRules(void) const;                               //!< Get argument rules
         
     };
